fix mkfs #1924

+++ lib/python/Components/Harddisk.py
@@ -457,7 +457,7 @@ class Harddisk:
 			task.args += ["-T", "largefile"]  # Between 16GB and 250GB: 1 i-node per megabyte.
 			big_o_options.append("sparse_super")
 		elif size > 2048:
-			task.args += ["-T", "largefile", "-N", str(size * 32)]  # Over 2GB: 32 i-nodes per megabyte.
+			task.args += ["-T", "largefile", "-N", str(int(size * 32))]  # Over 2GB: 32 i-nodes per megabyte.
 		task.args += ["-m0", "-O", ",".join(big_o_options), self.partitionPath("1")]
 		task = MountTask(job, self)
 		task.weighting = 3
